Move repositories to service
