--- docs/source/develop/domain/repository.rst
@@ -1,5 +1,0 @@
-Repositories
-============
-
-.. automodule:: purgatory.domain.repository
-   :members:
