+++ docs/source/develop/service/index.rst
@@ -4,6 +4,7 @@ Service
 .. toctree::
    :maxdepth: 2
 
+   repository
    unit_of_work
    messagebus
    handlers
