+++ docs/source/develop/service/repository.rst
@@ -1,0 +1,5 @@
+Repositories
+============
+
+.. automodule:: purgatory.service.repository
+   :members:
