--- src/purgatory/service/handlers/__init__.py
@@ -1,3 +1,0 @@
-from .circuitbreaker import register_circuit_breaker
-
-__all__ = ["register_circuit_breaker"]
