+++ src/purgatory/service/message_handlers.py
@@ -13,7 +13,9 @@ async def register_circuit_breaker(
     cmd: CreateCircuitBreaker, uow: AbstractUnitOfWork
 ) -> Context:
     """
-    Register circuit breaker in the repository when receiving the CreateCircuitBreaker command.
+    Register circuit breaker in the repository
+
+    when receiving the CreateCircuitBreaker command.
     """
     ret = Context(cmd.name, cmd.threshold, cmd.ttl)
     await uow.contexts.register(ret)
@@ -27,7 +29,9 @@ async def save_circuit_breaker_state(
     evt: ContextChanged, uow: AbstractUnitOfWork
 ) -> None:
     """
-    Save the circuit breaker state in the repository when receiving the ContextChanged event.
+    Save the circuit breaker state in the repository
+
+    when receiving the ContextChanged event.
     """
     await uow.contexts.update_state(evt.name, evt.state, evt.opened_at)
 
@@ -36,13 +40,17 @@ async def inc_circuit_breaker_failure(
     evt: CircuitBreakerFailed, uow: AbstractUnitOfWork
 ) -> None:
     """
-    Increment the number of failure in the repository when receiving the CircuitBreakerFailed event.
+    Increment the number of failure in the repository
+
+    when receiving the CircuitBreakerFailed event.
     """
     await uow.contexts.inc_failures(evt.name, evt.failure_count)
 
 
 async def reset_failure(evt: CircuitBreakerRecovered, uow: AbstractUnitOfWork) -> None:
     """
-    Reset the number of failure in the repository when receiving the CircuitBreakerRecovered event.
+    Reset the number of failure in the repository
+
+    when receiving the CircuitBreakerRecovered event.
     """
     await uow.contexts.reset_failure(evt.name)
